Add files via upload
Frequent Itemset Mining is a method for market basket analysis. In this project, the A-Priori algorithm is used to find all possible itemsets in textual tweets about the current Ukraine-Russian conflict. For this project, the Ukraine Conflict Twitter dataset is used. It is published on Kaggle and released under the CC-BY-SA 4.0 license. Dataset can be downloaded on https://www.kaggle.com/datasets/bwandowando/ukraine-russian-crisis-twitter-dataset-1-2-m-rows